repo-details removed, stargazers init
