+++ src/app.js
@@ -5,14 +5,8 @@ import 'whatwg-fetch';
 // Styles
 import 'app.scss';
 
-// Scripts (Components)
-import RepoDetails from 'scripts/RepoDetails';
-
 // Remove .no-js from html tag
 document.querySelector('html').classList.remove('no-js');
 
-// Initialize RepoDetails component
-const repoDetailsWrapper = document.getElementById('repo-details');
-if (repoDetailsWrapper) {
-  RepoDetails({ el: repoDetailsWrapper });
-}
+// Runnable (Not Exported) Components
+require('scripts/stargazers');
