+++ src/app.scss
@@ -9,4 +9,3 @@
 @import 'styles/layouts/footer';
 @import 'styles/layouts/header';
 @import 'styles/layouts/main';
-@import 'styles/layouts/repo-details';
