+++ src/index.html
@@ -15,7 +15,7 @@
   <header role="banner" class="header">
     <img src="assets/starbase.png" alt="starbase logo" class="header__image" />
     <h1 class="header__title">starbase</h1>
-    <a class="header__link link link--comment" href="https://github.com/bstaruk/starbase" title="learn more about starbase on GitHub">GitHub</a>
+    <a class="header__link link link--comment stargazers" href="https://github.com/bstaruk/starbase" title="learn more about starbase on GitHub" data-owner="bstaruk" data-repo="starbase">GitHub</a>
   </header>
 
   <!-- main content area -->
@@ -124,13 +124,6 @@ npm run build</code></pre>
     <p>starbase is open source and free software, so you may to do whatever you wish with it – commercially or
       personally. You can buy me a beer next time you’re in Boston, star the project and tell a friend, or you can erase
       all signs of origin and tell your coworkers that you made it yourself. It’s all good!</p>
-
-    <div id="repo-details" data-owner="bstaruk" data-repo="starbase">
-      <hr />
-      <h2>Who's using it</h2>
-      <p>Loading...</p>
-      <p class="repo-details__description">The information above was loaded asynchronously from GitHub using fetch. Fetch polyfills are included with the app so it'll still work perfectly with older browsers.</p>
-    </div>
   </main>
 
   <!-- footer -->
